requirements.txt
streamlit>=1.32.0
transformers==4.38.0
torch
sentence-transformers==2.2.2
faiss-cpu>=1.9.0
python-dotenv==1.0.0
pypdf==3.12.0
pdfplumber==0.8.0
tqdm==4.66.1
pillow>=10.2.0
